Add external TLS CA certificates
As the asio::ssl::context::set_default_verify_paths doesn't seem to work
on Android. The external database that comes with this commit was taken
from here:

https://curl.haxx.se/docs/caextract.html

Though it would be better to create our own such database by converting
Firefox's own database (not in pem format) into pem and use that.
Process of doing so is also described on that haxx.se page.