Revert "Decrease NDK version from r17b to r15c"
This reverts commit 3fd42f503abfebaff2703426948c158a507ac662.

NDK 15c is not supported any more, to build you have to use
o The NDK in this android.py file
o The version of clang ~/.mozbuild/clang
o The linker in ~/.mozbuild/clang
o The same version of rust as the Mozilla build server, currently 1.36.
  https://searchfox.org/mozilla-central/rev/c7e8bc4996f979e5876b33afae3de3b1ab4f3ae1/taskcluster/ci/toolchain/rust.yml#182

You can get the latest version of the first three by re-running ./mach bootstrap.