Bug 1641711 - Don't fail the build if javadoc fails. r=nalexander,snorp,esawin
`|mach lint -f treeherder|` expects lint to not fail ever, if a lint fails it's
just ignored. This means that when our javadoc errors out and fails the build
all lints are ignored.

To avoid that we add `failOnError = false` to the javadoc job so that we can
finish the build and actually report the errors.

This patch also adds gradle output `|mach lint|` to make it easier to
diagnose errors in the lint.

Differential Revision: https://phabricator.services.mozilla.com/D77373

(cherry picked from commit 519207a6a87eb299805036eb67cad8635fd5bbfd)