Turn "no Wi-Fi" dialog into a merely informative one.
Remove the possibility of quitting the app and service by choosing "Close",
which misguided users into thinking they were just closing the dialog.  Now
there should only be one way of quitting everything: via the notification
icon.

In the future we will extend the dialog to remember the decision and not show
up again.