Hint user about stopping app with mobile data (after losing Wi-Fi).
A previous commit removed from the "no Wi-Fi" dialog the mention of restarting
the app after losing connectivity, as the Ouinet client is now resilient
enough to recover from it.  However, as the dialog will also pop up even if
mobile data is not actually in use, we tell the user to check it out to inform
the decision on whether to close the app or not.

We should add checks to ensure that the dialog only pops up when using mobile
data, so that the message can be simplified.