TextView in private tab and change of color on normal tab background
As a way to test if this are the right places for modifications, a TextView was added to the layout believed to be in charge of displaying the content of private tabs.

Also, the background of what is believed to be a normal tab, was changed from transparent to red.